c++: Fix modules+contracts anonymous parameter friend decl issue
Previously, we were unconditionally saving the "new" parameter names in
function redeclarations in duplicate_decls for use in contracts parsing.
This causes issue when the "new" names are empty, and from a friend
decl, and we've already started parsing the body of the function. Now we
only save the argument names when contracts are enabled and we haven't
yet tried to parse the function's body.

2020-07-09  Jeff Chapman II  <jchapman@lock3software.com>

gcc/cp/
	* decl.c (duplicate_decls): Only save new argument names from
	function redecls when contracts are enabled and we haven't yet
	tried to parse the body of the function.

gcc/testsuite/
	* g++.dg/cpp2a/contracts-redecl1.C: New test case to ensure parm
	names are updated during non-defining redeclarations.
	* g++.dg/modules/contracts-tpl-friend-1_a.C:
	* g++.dg/modules/contracts-tpl-friend-1_b.C: New regression test.