Make dependent contract attributes work like other attributes.
Basically, ensure that contract attributes are instantiated by
tsubst_attribute, like any other attribute. Always indicate that
contract attributes are applied to the instantiaiton, not the template.
Sadly, we need to bypass a lookup for this to work.

This still crashes, but it seems like it's on the right path.