c++: Fix contract parm references when functions are finished
Previously, when a function was finished we were mistakenly replacing
the contract conditions on the function being finished rather than a
copy. This causes a number of issues because the original contract
conditions may be needed later (for instance if we're finishing a
base function and a later derived function inherits our contracts).

gcc/cp/
	* decl.c (finish_function_contracts): Manipulate a copy of the
	function contracts rather than its contracts directly.
	* search.c (remap_overrider_contracts): Cleanup.
	(check_final_overrider): Ditto.